+++ RELEASE_NOTES.md
@@ -1,3 +1,14 @@
+# Version 1.9
+
+## Features
+
+### Plate Solving
+- Added interface for ASTAP, the Astrometric STAcking Program, astrometric solver and FITS viewer as a plate solver
+
+## Improvements
+- If telescope is capable of reporting SideOfPier there will now be a new option to consider this for calculating the need for meridian flips
+- Added options to adjust downsample factor and maximum number of considered stars for ASTAP and local astrometry.net solvers
+
 # Version 1.8
 
 ## Features